backend/requirements.txt
fastapi
uvicorn
python-multipart
docling
sentence-transformers
qdrant-client
google-generativeai
python-dotenv
